moved functions to utils.py and removed app imports
