+++ Lumen-Server/app.py
@@ -15,6 +15,7 @@ import uuid
 import sys
 import os
 import env
+from utils import create_sensor_reading, get_historical_data_query
 
 port = sys.argv[1]
 mode = sys.argv[2]
@@ -76,85 +77,11 @@ def after_request(response):
     return response
 
 
-def _numeric_alarm_is_triggered(point_value, alarm_value, comparison_operator):
-    if comparison_operator == 'gt':
-        if point_value > alarm_value:
-            return True
-    elif comparison_operator == 'lt':
-        if point_value < alarm_value:
-            return True
-    elif comparison_operator == 'eq':
-        if point_value == alarm_value:
-            return True
-    return False
-
-
-def _discrete_alarm_is_triggered(point_value, alarm_value, comparison_operator):
-    if comparison_operator == 'eq':
-        if point_value == alarm_value:
-            return True
-    return False
-
-
-def _trigger_alarms(rtu_address, device_address, point_data):
-    devices = Device.select().where(
-        (Device.rtu_address == rtu_address) &
-        (Device.device_address == device_address)
-    )
-    for device in devices:
-        for alarm in device.alarms:
-            if alarm.data_type == 'number':
-                point_value = float(point_data[alarm.point])
-                alarm_value = float(alarm.value)
-                if _numeric_alarm_is_triggered(point_value, alarm_value, alarm.compare):
-                    alarm.is_triggered = 1
-                else:
-                    alarm.is_triggered = 0
-                alarm.save()
-            elif alarm.data_type == 'discrete':
-                point_value = point_data[alarm.point]
-                alarm_value = alarm.value
-                if _discrete_alarm_is_triggered(point_value, alarm_value, alarm.compare):
-                    alarm.is_triggered = 1
-                else:
-                    alarm.is_triggered = 0
-                alarm.save()
-
-
-def _trigger_alarms_for_all_devices():
-    devices = Device.select()
-    for device in devices:
-        point_data = json.loads(device.point_data)
-        if point_data == {}:
-            continue
-        _trigger_alarms(device.rtu_address,
-                        device.device_address,
-                        point_data)
-
-
-def _create_sensor_reading(data):
-
-    point_data_str = json.dumps(data['point_data'])
-    devices = Device.select().where(
-        (Device.rtu_address == data['rtu_address']) &
-        (Device.device_address == data['device_address'])
-    )
-    for device in devices:
-        device.point_data = point_data_str
-        device.save()
-
-    env.config['historical_database'].insert_one(data)
-
-    _trigger_alarms(data['rtu_address'],
-                    data['device_address'],
-                    data['point_data'])
-
-
 @app.route('/sensor-readings', methods=['POST'])
-def create_sensor_reading():
+def create_lab_sensor_reading():
     data = request.get_json()
     data['date'] = datetime.datetime.now()
-    _create_sensor_reading(data)
+    create_sensor_reading(data)
 
     return '', 201
 
@@ -179,43 +106,15 @@ def run_simulator_step():
     with open(filename) as f:
         step = json.load(f)
         for reading in step:
-            _create_sensor_reading(reading)
+            create_sensor_reading(reading)
 
         return '', 201
-
-
-def _get_historical_data_query(rtu_address, device_address):
-    query = {
-        'rtu_address': rtu_address,
-        'device_address': device_address,
-    }
-
-    for arg in request.args:
-        if len(request.args.get(arg)) > 0:
-            value = request.args.get(arg)
-            comparison, name = arg.split('_')
-            if name == 'date':
-                value = dateutil.parser.parse(value)
-                if comparison == 'gt':
-                    query['date'] = {'$gt': value}
-                elif comparison == 'lt':
-                    query['date'] = {'$lt': value}
-            else:
-                if comparison == 'eq':
-                    query[f'point_data.{name}'] = {'$eq': value}
-                elif comparison == 'gt':
-                    query[f'point_data.{name}'] = {'$gt': float(value)}
-                elif comparison == 'lt':
-                    query[f'point_data.{name}'] = {'$lt': float(value)}
-    return query
-
-
 
 
 @app.route('/historian-export/rtu-address/<rtu_address>/device-address/<device_address>', methods=['GET'])
 def get_historical_data_send_csv(rtu_address, device_address):
 
-    query = _get_historical_data_query(rtu_address, device_address)
+    query = get_historical_data_query(rtu_address, device_address, request)
     result = []
     for record in env.config['historical_database'].find(query):
         point_data = record['point_data']
